Update index.rst
Hi @andreap, thanks for asking for my opinion on this. My suggestions are in this new branch.  

A general comment: would the client have a name such as the ropentargets? Would you name it as opentargets-py, PyOpenTargets or something? 

I've mainly removed some adverbs (the fewer of them, the better - apparently. Check Eliseo's great tip http://www.hemingwayapp.com/)

Hope it's helpful.